Build a static framework; overhaul Couchbase.m
Added a 'Couchbase.framework' target to the Couchbase
project. The framework contains the static library code
(the same as produced by the existing bundle target), the
header file, and a 'CouchDB' subdirectory that contains
all the Erlang and other stuff needed at runtime.
A client app can now just add the framework to the
project, and a single copy-files build rule that copies
the 'CouchDB' resource dir into the output.

Overhauled the Couchbase class:
- Added a real API so that clients can configure its properties.
- Enabled -Werror in the Couchbase project.
- Reduced sleep times when Couchbase.m waits for the server to start;
  this reduces overall startup time.
- Gives up after 10 seconds if server doesn't start up.
- Other code cleanup of Couchbase.m.

Fixed a few compiler warnings in CouchDemo.

Changed the workspace file to override the default build location
with the 'DerivedData' subdirectory. This allows the build to work
correctly even if the user's specified a custom build location in
Xcode's prefs.
The build output (framework and apps) will be produced in:
DerivedData/Couchbase/Build/Products/Debug/