Server class updates out of date files on startup
The -install method in CouchbaseEmbeddedServer would only install the file if it didn't exist at the destination. That meant that those installed files would never get updated ever after the first app launch, which is very bad for upgrades.
Fixed it so that the installed file will be replaced if the source is newer.